Create separate GMS build flavour
