+++ .github/workflows/beta.yml
@@ -14,7 +14,7 @@ jobs:
     name: Create GH release and promote Play store internal to Beta
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v2      
+    - uses: actions/checkout@v2
     - name: Create release
       id: create_release
       uses: actions/create-release@v1
@@ -25,7 +25,7 @@ jobs:
         body_path: ./CHANGELOG.md
         release_name: ${{ github.ref }}
         draft: true
-        prerelease: true      
+        prerelease: true
     - name: set up JDK 11
       uses: actions/setup-java@v1
       with:
@@ -37,28 +37,37 @@ jobs:
           ~/.gradle/wrapper
         key: ${{ runner.os }}-gradle-${{ hashFiles('**/*.gradle*') }}
         restore-keys: |
-          ${{ runner.os }}-gradle-        
+          ${{ runner.os }}-gradle-
     - name: Decrypt secrets
       env:
-        GOOGLE_CLOUD_SERVICE_ACCOUNT_CREDENTIALS: ${{ secrets.GOOGLE_CLOUD_SERVICE_ACCOUNT_CREDENTIALS }}          
+        GOOGLE_CLOUD_SERVICE_ACCOUNT_CREDENTIALS: ${{ secrets.GOOGLE_CLOUD_SERVICE_ACCOUNT_CREDENTIALS }}
         KEYSTORE_BASE64: ${{ secrets.KEYSTORE_BASE64 }}
       run: |
         echo -n $GOOGLE_CLOUD_SERVICE_ACCOUNT_CREDENTIALS > project/app/owntracks-android-7a8e1517bde3.json
         echo -n $KEYSTORE_BASE64 | base64 -d > project/owntracks.release.keystore.jks
     - name: Build APK
-      run: ./gradlew assemble
+      run: ./gradlew assembleRelease
       working-directory: project
       env:
         KEYSTORE_PASSPHRASE: ${{ secrets.KEYSTORE_PASSPHRASE }}
         ORG_GRADLE_PROJECT_google_maps_api_key: ${{ secrets.GOOGLE_MAPS_API_KEY }}
-    - name: Upload Release Asset
+    - name: Upload GMS Release Asset
       uses: actions/upload-release-asset@v1
       env:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
       with:
         upload_url: ${{ steps.create_release.outputs.upload_url }} # This pulls from the CREATE RELEASE step above, referencing it's ID to get its outputs object, which include a `upload_url`. See this blog post for more info: https://jasonet.co/posts/new-features-of-github-actions/#passing-data-to-future-steps 
-        asset_path: ./project/app/build/outputs/apk/release/app-release.apk
-        asset_name: owntracks-release.apk
+        asset_path: ./project/app/build/outputs/apk/gms/release/app-gms-release.apk
+        asset_name: owntracks-release-gms.apk
+        asset_content_type: application/vnd.android.package-archive
+    - name: Upload OSS Release Asset
+      uses: actions/upload-release-asset@v1
+      env:
+        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      with:
+        upload_url: ${{ steps.create_release.outputs.upload_url }} # This pulls from the CREATE RELEASE step above, referencing it's ID to get its outputs object, which include a `upload_url`. See this blog post for more info: https://jasonet.co/posts/new-features-of-github-actions/#passing-data-to-future-steps 
+        asset_path: ./project/app/build/outputs/apk/oss/release/app-oss-release.apk
+        asset_name: owntracks-release-oss.apk
         asset_content_type: application/vnd.android.package-archive
     - name: Promote play store beta from internal
       run: ./gradlew promoteArtifact --from-track internal --promote-track beta --release-status completed
